Rearranging conditionals for setting 'forward' buttons - Fixed: Datasets now render in ColumnEdit immediately after cohort selection; Select dataset view retains its previous view after clicking back from Select Data Splice view - Added: Select dataset view retains user state, e.g. collapsed accordion panel stays as-is - Fixed: Reinstated 'Samples in' section in AppControls; refactored how datasets are being sorted when grouping and presenting - Changed: moved all sorting operations into existing 'sortDatasets' helper function
